src/Messaging/MemberEventDetails.ts
import { Document } from "../Document/Document";

import { ObjectSchema } from "../ObjectSchema/ObjectSchema";

import { Resource } from "../Resource/Resource";

import { C } from "../Vocabularies/C";


export interface MemberEventDetails extends Resource {
	members:Document[];
}


export interface MemberEventDetailsFactory {
	SCHEMA:ObjectSchema;
}

const SCHEMA:ObjectSchema = {
	"members": {
		"@id": C.member,
		"@type": "@id",
		"@container": "@set",
	},
};

export const MemberEventDetails:MemberEventDetailsFactory = {
	SCHEMA,
};
